requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
timm==0.9.12
pillow>=10.0.0
jinja2==3.1.2
